[REF] 모호한 class name 변경
