+++ src/main/java/racingcar/view/InputView.java
@@ -1,6 +1,6 @@
 package racingcar.view;
 
-import racingcar.utils.Validator;
+import racingcar.utils.CarValidator;
 
 import java.util.Scanner;
 
@@ -11,7 +11,7 @@ public class InputView {
         System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
         try {
             String userCarNames = SCANNER.nextLine();
-            Validator.checkCarNames(userCarNames);
+            CarValidator.checkCarNames(userCarNames);
             return userCarNames;
         } catch (IllegalArgumentException e) {
             return getUserCarNames();
@@ -22,7 +22,7 @@ public class InputView {
         System.out.println("시도할 회수는 몇회인가요?");
         try {
             int numberOfGame = Integer.parseInt(SCANNER.nextLine());
-            Validator.checkNumberOfGame(numberOfGame);
+            CarValidator.checkNumberOfGame(numberOfGame);
             return numberOfGame;
         } catch (IllegalArgumentException e) {
             return getNumberOfGame();
